最新

+++ Assets/Scripts/ScriptableObject/Magic/FlashSO.cs
@@ -8,9 +8,9 @@ public class FlashSO : MagicBaseSO
     public override void Execute(Battler user, Battler target)
     {
         base.Execute(user, target);
-        int successRate = user.men - target.men;
+        float successRate = user.men - target.men;
         if (successRate < 5) { successRate = 5; }
-        int rundomNumber = Random.Range(0, 101);
+        float rundomNumber = Random.Range(0, 101);
 
         if( rundomNumber <= successRate)
         {
